python-src/ver1.0.1/requirements.txt
# Function dependencies, for example:
# package>=version
line-bot-sdk
flask
google.cloud.logging
